// src/types/database.types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "12.2.12 (cd3cf9e)"
  }
  public: {
    Tables: {
// ...
      patients: {
        Row: {
          age: number | null
          avatar: string | null
          created_at: string | null
          gender: Database["public"]["Enums"]["gender_type"] | null
          id: string
          name: string
          priority: Database["public"]["Enums"]["priority_level"] | null
          requested_time: string | null
          serial_no: number
          symptoms: string | null
          updated_at: string | null
          aadhaar_number: string | null
          email: string | null
        }
        Insert: {
          age?: number | null
          avatar?: string | null
          created_at?: string | null
          gender?: Database["public"]["Enums"]["gender_type"] | null
          id?: string
          name: string
          priority?: Database["public"]["Enums"]["priority_level"] | null
          requested_time?: string | null
          serial_no?: number
          symptoms?: string | null
          updated_at?: string | null
          aadhaar_number?: string | null
          email?: string | null
        }
        Update: {
          age?: number | null
          avatar?: string | null
          created_at?: string | null
          gender?: Database["public"]["Enums"]["gender_type"] | null
          id?: string
          name?: string
          priority?: Database["public"]["Enums"]["priority_level"] | null
          requested_time?: string | null
          serial_no?: number
          symptoms?: string | null
          updated_at?: string | null
          aadhaar_number?: string | null
          email?: string | null
        }
        Relationships: []
      }
// ...
    }
// ...
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type TablesInsert<
  DefaultSchemaTableNameOrOptions extends
    | keyof DefaultSchema["Tables"]
    | { schema: keyof DatabaseWithoutInternals },
  TableName extends DefaultSchemaTableNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"]
    : never = never,
> = DefaultSchemaTableNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"][TableName] extends {
      Insert: infer I
    }
    ? I
    : never
  : DefaultSchemaTableNameOrOptions extends keyof DefaultSchema["Tables"]
    ? DefaultSchema["Tables"][DefaultSchemaTableNameOrOptions] extends {
        Insert: infer I
      }
      ? I
      : never
    : never